tweaks to general math

+++ headless/src/main/java/sh/miles/algidle/computer/Equation.java
@@ -1,6 +1,7 @@
 package sh.miles.algidle.computer;
 
+import java.math.BigInteger;
 import java.util.function.Function;
 
-public record Equation(String name, Function<Integer, Integer> time) {
+public record Equation(String name, Function<BigInteger, BigInteger> time) {
 }
